+++ api-colegio/src/main/java/com/apolo/modulos/acudiente/service/AcudienteService.java
@@ -72,7 +72,7 @@ public class AcudienteService implements IAcudienteService {
 
 
     @Override
-    public Optional<Acudiente> findById(Integer id) {
+    public Optional<Acudiente> findById(Long id) {
         Optional<Acudiente> acudiente = acudienteRepository.findById(id);
         if (!acudiente.isPresent())
             throw new ObjetoNoEncontradoException("No se existe un acudiente con el usuario id: " + id);
