+++ api-colegio/src/main/java/com/apolo/modulos/roles/repository/RolUsuarioRepository.java
@@ -1,11 +1,9 @@
 package com.apolo.modulos.roles.repository;
 
-import com.apolo.modulos.roles.repository.RolUsuario;
-
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RolUsuarioRepository extends JpaRepository<RolUsuario, Integer> {
+public interface RolUsuarioRepository extends JpaRepository<RolUsuario, Long> {
 
 }
